requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
executing==2.2.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.56.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pika==1.3.2
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg==3.2.6
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pyzmq==26.4.0
requests==2.31.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.1
wcwidth==0.2.13

# Multi-agent & LLM orchestration
langchain==0.3.25
langgraph==0.4.1
langgraph-supervisor==0.0.20
